Reduce building request range
