+++ client/src/main/kotlin/dev/voroscsoki/stratopolis/client/user_interface/UtilInput.kt
@@ -4,6 +4,7 @@ import com.badlogic.gdx.Input
 import com.badlogic.gdx.InputAdapter
 import dev.voroscsoki.stratopolis.client.Main
 import dev.voroscsoki.stratopolis.client.graphics.MainScene
+import dev.voroscsoki.stratopolis.client.graphics.toWorldCoords
 import dev.voroscsoki.stratopolis.common.networking.BuildingRequest
 import dev.voroscsoki.stratopolis.common.networking.NodeRequest
 import dev.voroscsoki.stratopolis.common.networking.SimulationStartRequest
@@ -21,8 +22,8 @@ class UtilInput(val scene: MainScene) : InputAdapter() {
         }
         if (keycode == 132) {
             runBlocking {
-                val source = scene.cam.position.let { Vec3(it.x, it.y, it.z) }
-                Main.socket.sendSocketMessage(BuildingRequest(source, 0.00001f)) }
+                val source = scene.cam.position?.toWorldCoords(scene.baselineCoord)!!.copy(y = 0f)
+                Main.socket.sendSocketMessage(BuildingRequest(source, 0.1f)) }
         }
         if (keycode == 133) {
             runBlocking {
